Extend querySelectorAll to pierce Shadow DOM
This is useful for pages where Custom Elements are present and authors wish to use icons inside of a custom element's shadow tree.

This only supports entire selectors within a document (or shadow) boundary, and does not support matching part of a selector outside of a shadow tree, while matching the rest of it inside that shadow tree. This limited functionality was chosen as the simplest way to extend the  package to support  across shadow boundaries while only having to call it once.